Refactor document and component structure
