+++ HtmlForgeX/Containers/DataTables/DataTablesOptions.cs
@@ -49,60 +49,3 @@ public class DataTablesOptions
     public string? Dom { get; set; }
 }
 
-public class DataTablesLanguage
-{
-    [JsonPropertyName("search")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Search { get; set; }
-
-    [JsonPropertyName("lengthMenu")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? LengthMenu { get; set; }
-
-    [JsonPropertyName("info")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Info { get; set; }
-
-    [JsonPropertyName("infoEmpty")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? InfoEmpty { get; set; }
-
-    [JsonPropertyName("infoFiltered")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? InfoFiltered { get; set; }
-
-    [JsonPropertyName("loadingRecords")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? LoadingRecords { get; set; }
-
-    [JsonPropertyName("processing")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Processing { get; set; }
-
-    [JsonPropertyName("zeroRecords")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? ZeroRecords { get; set; }
-
-    [JsonPropertyName("paginate")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public DataTablesPaginate? Paginate { get; set; }
-}
-
-public class DataTablesPaginate
-{
-    [JsonPropertyName("first")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? First { get; set; }
-
-    [JsonPropertyName("last")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Last { get; set; }
-
-    [JsonPropertyName("next")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Next { get; set; }
-
-    [JsonPropertyName("previous")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Previous { get; set; }
-}
