+++ HtmlForgeX/Containers/DataTables/DataTablesPaginate.cs
@@ -1,0 +1,23 @@
+using System.Text.Json.Serialization;
+
+namespace HtmlForgeX;
+
+public class DataTablesPaginate
+{
+    [JsonPropertyName("first")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public string? First { get; set; }
+
+    [JsonPropertyName("last")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public string? Last { get; set; }
+
+    [JsonPropertyName("next")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public string? Next { get; set; }
+
+    [JsonPropertyName("previous")]
+    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
+    public string? Previous { get; set; }
+}
+
