+++ HtmlForgeX/Containers/Tabler/Icons/TablerIcon.cs
@@ -5,20 +5,6 @@ using HtmlForgeX.Extensions;
 
 namespace HtmlForgeX;
 
-/// <summary>
-/// Wrapper for raw HTML content that should not be encoded
-/// </summary>
-internal class RawHtml {
-    public string Content { get; }
-
-    public RawHtml(string content) {
-        Content = content;
-    }
-
-    public override string ToString() {
-        return Content;
-    }
-}
 
 /// <summary>
 /// Represents an SVG icon with full styling and positioning flexibility
