+++ HtmlForgeX/Containers/Tabler/Icons/TablerIconType.cs
